Make 'items' the canonical term and 'members' the alias
I mostly used 'members' b/c I saw it in Set::NestedGroups
and it made some sense, but I wasn't fond of it.

Outside of the context of groups it doesn't make sense
(like calling the members() method to get all known members...
if they're not in groups they're not members of anything).

So rename the variables and methods for consistency and clarity.